requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
requests==2.32.3
python-dotenv==1.0.1
langchain==0.3.0
openai==1.45.1
faiss-cpu==1.8.0
